+++ samila/functions.py
@@ -412,27 +412,6 @@ def generate_params_filter(
     g.seed, g.start, g.step, g.stop = seed, start, step, stop
 
 
-def fill_data(g, point):
-    """
-    Fill data with functions in given points.
-
-    :param g: generative image instance
-    :type g: GenerativeImage
-    :param point: given point
-    :type point: tuple
-    :return: false if some exception occurred
-    """
-    random.seed(g.seed)
-    try:
-        data1_ = g.function1(point[0], point[1]).real
-        data2_ = g.function2(point[0], point[1]).real
-    except Exception:
-        return False
-    g.data1.append(data1_)
-    g.data2.append(data2_)
-    return True
-
-
 def save_params_filter(g, depth=None):
     """
     Filter save_image method parameters.
@@ -480,6 +459,7 @@ def _GI_initializer(g, function1, function2):
     g.alpha = DEFAULT_ALPHA
     g.linewidth = DEFAULT_LINEWIDTH
     g.depth = DEFAULT_DEPTH
+    g.missed_points_number = 0
 
 
 def nft_storage_upload(api_key, data):