+++ CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Marker` enum
 ### Changed
 - `marker` parameter added to `plot` method
+- `generate` method optimized
 - Test system modified
 - `Python 3.11` added to `test.yml`
+### Removed
+- `fill_data` function
 ## [0.9] - 2022-09-28
 ### Added
 - Anaconda workflow